Warpbox distorts rotated bounding boxes
Because the width and height calculated in WarpBox are taken from cv2.boundingRect, the non-rotated bbox width and height are used, which distorts the images. The proposed change calculates the length of the rotated rectangle width and height and warps it accordingly.
Example of previous output and new output: https://imgur.com/a/us3Dxq9 
The code assumes the points are ordered correctly, which should be guaranteed by get_rotated_box.